add features

+++ src/main/kotlin/ru/remsoftware/game/money/boosters/Booster.kt
@@ -1,12 +1,9 @@
 package ru.remsoftware.game.money.boosters
 
-import org.bukkit.boss.BossBar
-import kotlin.time.Duration
-
 data class Booster(
     val activationTime: Long,
-    val duration: Long,
-    var remainingTime: Long,
+    val duration: Int,
+    var remainingTime: Int,
     val local: Boolean,
     val activator: String,
 )
